Add files via upload
This file contains the code and assets for a single-page website, designed to provide a concise and focused user experience. The website is built using Html, Css and offers a convenient way for reviewers to access my contact information and get in touch with me. The website features a simple and easy-to-navigate design, ensuring that the most important information is prominently displayed and easily accessible.